Add quantity

+++ x/nft/internal/keeper/nft.go
@@ -5,8 +5,9 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/modules/incubator/nft/exported"
-	"github.com/cosmos/modules/incubator/nft/types"
+
+	"bitbucket.org/decimalteam/go-node/x/nft/exported"
+	"bitbucket.org/decimalteam/go-node/x/nft/internal/types"
 )
 
 // IsNFT returns whether an NFT exists
@@ -29,32 +30,6 @@ func (k Keeper) GetNFT(ctx sdk.Context, denom, id string) (nft exported.NFT, err
 	return nft, err
 }
 
-// UpdateNFT updates an already existing NFTs
-func (k Keeper) UpdateNFT(ctx sdk.Context, denom string, nft exported.NFT) (err error) {
-	collection, found := k.GetCollection(ctx, denom)
-	if !found {
-		return sdkerrors.Wrap(types.ErrUnknownCollection, fmt.Sprintf("collection #%s doesn't exist", denom))
-	}
-	oldNFT, err := collection.GetNFT(nft.GetID())
-	if err != nil {
-		return err
-	}
-	// if the owner changed then update the owners KVStore too
-	if !oldNFT.GetOwner().Equals(nft.GetOwner()) {
-		err = k.SwapOwners(ctx, denom, nft.GetID(), oldNFT.GetOwner(), nft.GetOwner())
-		if err != nil {
-			return err
-		}
-	}
-	collection, err = collection.UpdateNFT(nft)
-
-	if err != nil {
-		return err
-	}
-	k.SetCollection(ctx, denom, collection)
-	return nil
-}
-
 // MintNFT mints an NFT and manages that NFTs existence within Collections and Owners
 func (k Keeper) MintNFT(ctx sdk.Context, denom string, nft exported.NFT) (err error) {
 	collection, found := k.GetCollection(ctx, denom)
@@ -68,14 +43,14 @@ func (k Keeper) MintNFT(ctx sdk.Context, denom string, nft exported.NFT) (err er
 	}
 	k.SetCollection(ctx, denom, collection)
 
-	ownerIDCollection, _ := k.GetOwnerByDenom(ctx, nft.GetOwner(), denom)
+	ownerIDCollection, _ := k.GetOwnerByDenom(ctx, nft.GetCreator(), denom)
 	ownerIDCollection = ownerIDCollection.AddID(nft.GetID())
-	k.SetOwnerByDenom(ctx, nft.GetOwner(), denom, ownerIDCollection.IDs)
+	k.SetOwnerByDenom(ctx, nft.GetCreator(), denom, ownerIDCollection.IDs)
 	return
 }
 
 // DeleteNFT deletes an existing NFT from store
-func (k Keeper) DeleteNFT(ctx sdk.Context, denom, id string) (err error) {
+func (k Keeper) DeleteNFT(ctx sdk.Context, denom, id string, quantity sdk.Int) (err error) {
 	collection, found := k.GetCollection(ctx, denom)
 	if !found {
 		return sdkerrors.Wrap(types.ErrUnknownCollection, fmt.Sprintf("collection of %s doesn't exist", denom))
@@ -84,22 +59,27 @@ func (k Keeper) DeleteNFT(ctx sdk.Context, denom, id string) (err error) {
 	if err != nil {
 		return err
 	}
-	ownerIDCollection, found := k.GetOwnerByDenom(ctx, nft.GetOwner(), denom)
+	ownerIDCollection, found := k.GetOwnerByDenom(ctx, nft.GetCreator(), denom)
 	if !found {
 		return sdkerrors.Wrap(types.ErrUnknownCollection,
-			fmt.Sprintf("id collection #%s doesn't exist for owner %s", denom, nft.GetOwner()),
+			fmt.Sprintf("id collection #%s doesn't exist for owner %s", denom, nft.GetCreator()),
 		)
 	}
-	ownerIDCollection, err = ownerIDCollection.DeleteID(nft.GetID())
-	if err != nil {
-		return err
-	}
-	k.SetOwnerByDenom(ctx, nft.GetOwner(), denom, ownerIDCollection.IDs)
 
-	collection, err = collection.DeleteNFT(nft)
-	if err != nil {
-		return err
+	if quantity.GT(nft.GetOwners().GetOwner(nft.GetCreator()).GetQuantity()) {
+		return sdkerrors.Wrap(types.ErrNotAllowedBurn,
+			fmt.Sprintf("owner %s has only %s tokens", nft.GetCreator(), nft.GetOwners().GetOwner(nft.GetCreator()).GetQuantity().String()))
 	}
+
+	if quantity.Equal(nft.GetOwners().GetOwner(nft.GetCreator()).GetQuantity()) {
+		ownerIDCollection, err = ownerIDCollection.DeleteID(nft.GetID())
+		if err != nil {
+			return err
+		}
+		k.SetOwnerByDenom(ctx, nft.GetCreator(), denom, ownerIDCollection.IDs)
+	}
+
+	nft.GetOwners().GetOwner(nft.GetCreator()).SetQuantity(nft.GetOwners().GetOwner(nft.GetCreator()).GetQuantity().Sub(quantity))
 
 	k.SetCollection(ctx, denom, collection)
